requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
numpy>=1.26.4
tenacity>=8.2.3
